requirements.txt
pytest~=6.2.5
certifi==2021.10.8
charset-normalizer==2.0.7
greenlet==1.1.2
idna==3.2
password==0.2
psycopg==3.0
PyMySQL==1.0.2
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
vk==2.0.2
vk-api==11.9.4